requirements.txt
# Original dependencies
fastapi
uvicorn
numpy
nltk
python-multipart
python-dotenv
streamlit

# Local models (replaces Mistral API)
sentence-transformers
torch
torchvision

# Local PDF parsing (replaces Mistral OCR)
pdfplumber
PyMuPDF

# Optimization tools
numba
faiss-cpu

# Profiling
memory-profiler
line-profiler
matplotlib
pandas
seaborn

# Optional: GPU support (uncomment if GPU available)
# cupy-cuda12x
# faiss-gpu
